// src/model/Her2FishResults.ts
export type Her2FishResults = {
    specimen: string;
    source: string;
    tissue_id: string;
    order_id: string;
    ordered_date: string;
    ihc:string;
    num_nuclei: number;
    num_her2: number;
    num_cep17: number;
    completed:boolean;
    completed_by:string;
    completed_datetime:string;
    finalized:boolean;
    finalized_by:string;
    finalized_datetime:string;
}